+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Added: `forceDiscrepancy` option for `openCookieClickerPage`.
 - Added: In `CCSave.fromObject`,
     the `.achievements` attribute may be just the string `'all'`.
 - Fixed: Pretty-printer of `npx cookie-connoisseur native-to-json`